requirements.txt
# Core dependencies
earthengine-api>=0.1.374
geemap>=0.33.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
xarray>=2023.1.0
rasterio>=1.3.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=1.7.0
ruptures>=1.1.8

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API and Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
streamlit>=1.25.0
streamlit-folium>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pydantic>=2.0.0
